+++ diesel/src/migration/mod.rs
@@ -4,18 +4,27 @@ mod errors;
 pub use self::errors::{MigrationError, RunMigrationsError};
 
 use connection::SimpleConnection;
+use std::borrow::Cow;
 use std::path::Path;
 
 /// Represents a migration that interacts with diesel
 pub trait Migration {
     /// Get the migration version
     fn version(&self) -> &str;
+
     /// Apply this migration
     fn run(&self, conn: &SimpleConnection) -> Result<(), RunMigrationsError>;
+
     /// Revert this migration
     fn revert(&self, conn: &SimpleConnection) -> Result<(), RunMigrationsError>;
+
     /// Get the migration file path
     fn file_path(&self) -> Option<&Path> {
+        None
+    }
+
+    /// Get the metadata associated with this migration, if any
+    fn metadata(&self) -> Option<&Metadata> {
         None
     }
 }
@@ -32,8 +41,13 @@ impl Migration for Box<Migration> {
     fn revert(&self, conn: &SimpleConnection) -> Result<(), RunMigrationsError> {
         (&**self).revert(conn)
     }
+
     fn file_path(&self) -> Option<&Path> {
         (&**self).file_path()
+    }
+
+    fn metadata(&self) -> Option<&Metadata> {
+        (&**self).metadata()
     }
 }
 
@@ -49,7 +63,27 @@ impl<'a> Migration for &'a Migration {
     fn revert(&self, conn: &SimpleConnection) -> Result<(), RunMigrationsError> {
         (&**self).revert(conn)
     }
+
     fn file_path(&self) -> Option<&Path> {
         (&**self).file_path()
     }
+
+    fn metadata(&self) -> Option<&Metadata> {
+        (&**self).metadata()
+    }
+}
+
+/// Represents metadata associated with a migration.
+///
+/// The format of a migration's metadata is dependent on the migration format
+/// being used.
+///
+/// For Diesel's built in SQL file migrations, metadata is stored in a file
+/// called `metadata.toml`. Diesel looks for a single key, `run_in_transaction`.
+/// By default, all migrations are run in a transaction on SQLite and
+/// PostgreSQL. This behavior can be disabled for a single migration by setting
+/// this to `false`.
+pub trait Metadata {
+    /// Get the metadata at the given key, if present
+    fn get(&self, key: &str) -> Option<Cow<str>>;
 }